Se arreglaron los metodos de la tabla usuarios y se modifico el metodo de eliminar usuarios

+++ classes/user.php
@@ -158,12 +158,16 @@ class user {
         $select_photo = $object -> _db -> prepare("SELECT nombre_foto, foto_identificador FROM usuarios WHERE id=:iduser");
         $select_photo -> execute(array(':iduser' => $id));
         $fetch_select_photo = $select_photo -> fetch(PDO::FETCH_OBJ);
-        $directory = __DIR__ . "/../src/img/imgs_uploaded/";
-        $path = __DIR__ . "/../src/img/imgs_uploaded/".$fetch_select_photo -> foto_identificador;
-        if(!file_exists($path)){
-            $crud->delete('usuarios', 'id=:iduser', ['iduser' => $id]);
+        if($fetch_select_photo -> nombre_foto != null && $fetch_select_photo -> foto_identificador != null){
+            $directory = __DIR__ . "/../src/img/imgs_uploaded/";
+            $path = __DIR__ . "/../src/img/imgs_uploaded/".$fetch_select_photo -> foto_identificador;
+            if(!file_exists($path)){
+                $crud->delete('usuarios', 'id=:iduser', ['iduser' => $id]);
+            }else{
+                unlink($directory.$fetch_select_photo -> foto_identificador);
+                $crud->delete('usuarios', 'id=:iduser', ['iduser' => $id]);
+            }
         }else{
-            unlink($directory.$fetch_select_photo -> foto_identificador);
             $crud->delete('usuarios', 'id=:iduser', ['iduser' => $id]);
         }
 	}
